fix: resovle PR comments and update to simplify the code structure
