+++ src/services/larksuite/appointment/appointment-service.ts
@@ -8,17 +8,8 @@ import utc from "dayjs/plugin/utc.js";
 import timezone from "dayjs/plugin/timezone.js";
 import LarksuiteService from "src/services/larksuite/lark";
 import { TIMEZONE_VIETNAM } from "src/constants";
-import normalizePhoneNumber from "services/utils/normalize-phone-number";
-
-dayjs.extend(utc);
-dayjs.extend(timezone);
-
-const GENDER_MAP: Record<string, string> = {
-  "nam": "Male",
-  "nữ": "Female",
-  "lgbt": "LGBT"
-};
-
+import { fetchLeadInfoByPhoneNumber } from "frappe/lead";
+import { mapLarkToFrappe } from "frappe/utils/utils-lark";
 export default class AppointmentService {
   env: any;
   db: PrismaClient;
@@ -46,7 +37,7 @@ export default class AppointmentService {
     });
   }
 
-  async syncLarkRecord(appToken: string, tableId: string, recordId: string): Promise<ILarksuiteAppointment> {
+  private async syncLarkRecord(appToken: string, tableId: string, recordId: string): Promise<ILarksuiteAppointment> {
     // @ts-expect-error This RecordService was written in javascript so we can not define the type for it
     const record = await RecordService.getLarksuiteRecord({
       env: this.env,
@@ -60,20 +51,21 @@ export default class AppointmentService {
     }
 
     const rawFields = (record.fields || {}) as LarksuiteAppointmentRawFields;
+    console.log('rawFields', rawFields);
     const fields: LarksuiteAppointmentParsedFields = {
       store: rawFields["Cửa hàng"]?.[0] || "",
       name: rawFields["Tên khách hàng/ facebook"] || "",
       phone_number: rawFields["Số điện thoại"] || "",
       gender: rawFields["Giới tính"] || "",
-      product_images: rawFields["Hình ảnh sản phẩm (nếu có)"] || null,
-      note: rawFields["Lưu ý đặc biệt"] || null,
+      product_images: rawFields["Hình ảnh sản phẩm (nếu có)"],
+      note: rawFields["Lưu ý đặc biệt"],
       date_time: rawFields["Ngày khách dự kiến tới CH"] ? new Date(rawFields["Ngày khách dự kiến tới CH"]) : null,
-      conversation_greeting: rawFields["Nội dung đón tiếp tại cửa hàng"] || null,
-      customer_response: rawFields["Offlie Phản hồi"] || null,
-      main_sales: rawFields["Sale chính"] || null,
-      offline_sales: rawFields["Sale Offline tiếp nhận"] || null,
-      status: rawFields["Trạng thái đơn hàng"] || null,
-      policy: rawFields["Chính sách thu mua thu đổi"] || null
+      conversation_greeting: rawFields["Nội dung đón tiếp tại cửa hàng"],
+      customer_response: rawFields["Offlie Phản hồi"],
+      main_sales: rawFields["Sale chính"],
+      offline_sales: rawFields["Sale Offline tiếp nhận"],
+      status: rawFields["Trạng thái đơn hàng"],
+      policy: rawFields["Chính sách thu mua thu đổi"]
     };
 
     const data: ILarksuiteAppointment = {
@@ -87,7 +79,7 @@ export default class AppointmentService {
     return data;
   }
 
-  async getDocumentAttachments(doctype: string, docname: string): Promise<IFrapperAttachment[]> {
+  private async getDocumentAttachments(doctype: string, docname: string): Promise<IFrapperAttachment[]> {
     try {
       const attachments = await this.frappeClient.getList("File", {
         fields: ["name", "file_name", "file_url", "is_private"],
@@ -103,7 +95,7 @@ export default class AppointmentService {
     }
   }
 
-  async downloadFileAndUploadFrappe(attachments: LarksuiteAttachment[], docname: string) {
+  private async downloadFileAndUploadFrappe(attachments: LarksuiteAttachment[], docname: string) {
     try {
       if (!attachments || attachments.length === 0) return;
       const accessToken = await LarksuiteService.getTenantAccessToken(this.env);
@@ -144,96 +136,15 @@ export default class AppointmentService {
     }
   }
 
-  async fetchLeadInfoByPhoneNumber(phoneNumber: string) {
-    try {
-      const lead: IFrappeLead[] = await this.frappeClient.getList("Lead", {
-        fields: ["name", "first_name", "budget_lead", "proposed_budget", "phone", "email_id"],
-        filters: {
-          phone: normalizePhoneNumber(phoneNumber)
-        }
-      });
-      return lead;
-    } catch (error) {
-      console.warn("Error fetching lead info:", error);
-      return null;
-    }
-  }
-
-  async getAllSalesPersons(employeeEmails: string[], names: string[]): Promise<IFrappeSalesPerson[]> {
-    if ((!employeeEmails || employeeEmails.length === 0) && (!names || names.length === 0)) {
-      return [];
-    }
-
-    try {
-      const or_filters: any[] = [];
-      if (employeeEmails && employeeEmails.length > 0) {
-        or_filters.push(["employee_email", "in", employeeEmails]);
-      }
-      if (names && names.length > 0) {
-        or_filters.push(["sales_person_name", "in", names]);
-      }
-
-      const salesPersons = await this.frappeClient.getList("Sales Person", {
-        fields: ["name", "sales_person_name", "employee_email"],
-        or_filters: or_filters,
-        limit_page_length: 100
-      });
-
-      return salesPersons || [];
-    } catch (error) {
-      console.warn("Error fetching Sales Persons:", error);
-      return [];
-    }
-  }
-
-  private async mapLarkToFrappe(dataRequest: ILarksuiteAppointment, lead?: IFrappeLead) {
-    const genderLower = dataRequest.gender?.toLowerCase() || "";
-    const gender = GENDER_MAP[genderLower] || "Male";
-
-    const scheduledTime = dataRequest.date_time
-      ? dayjs(dataRequest.date_time).tz(TIMEZONE_VIETNAM).format("YYYY-MM-DD HH:mm:ss")
-      : undefined;
-
-    const allEmails = [...(dataRequest.main_sales || []), ...(dataRequest.offline_sales || [])].map(s => s.email).filter(Boolean);
-    const allNames = [...(dataRequest.main_sales || []), ...(dataRequest.offline_sales || [])].map(s => s.name).filter(Boolean);
-    const salesPersons = (allEmails.length > 0 || allNames.length > 0) ? await this.getAllSalesPersons(allEmails, allNames) : [];
-
-    const mapSalesPerson = (person: LarksuiteSalePerson) => {
-      const matched = salesPersons.find((sp: IFrappeSalesPerson) =>
-        (sp.employee_email && sp.employee_email === person.email) ||
-        (sp.sales_person_name && sp.sales_person_name === person.name)
-      );
-      if (matched) {
-        return { sales_person: matched.name };
-      }
-      return null;
-    };
-
-    const main_sales = (dataRequest.main_sales || []).map(mapSalesPerson);
-    const offline_sales = (dataRequest.offline_sales || []).map(mapSalesPerson);
-
-    return {
-      customer_name: dataRequest.name || "Khách hàng",
-      customer_phone_number: dataRequest.phone_number,
-      scheduled_time: scheduledTime,
-      status: dataRequest.status || "Khách hẹn đến cửa hàng",
-      store: dataRequest.store,
-      gender: gender,
-      conversation_greeting: dataRequest.conversation_greeting,
-      customer_response: dataRequest.customer_response,
-      notes: dataRequest.note,
-      record_id: dataRequest.record_id,
-      policy: dataRequest.policy,
-      lead: lead ? lead.name : undefined,
-      estimated_budget: lead ? lead.budget_lead : undefined,
-      range_estimated_budget: lead ? lead.proposed_budget : undefined,
-      main_sales: main_sales.filter(Boolean),
-      offline_sales: offline_sales.filter(Boolean)
-    };
+  async createOrUpdateAppointment(appToken: string, tableId: string, recordId: string) {
+    const record = await this.syncLarkRecord(appToken, tableId, recordId);
+    const lead = await fetchLeadInfoByPhoneNumber(this.frappeClient, record.phone_number);
+    const erpAppointment = await this.upsertERPAppointment(record, lead);
+    return erpAppointment
   }
 
   async createNewERPAppointment(dataRequest: ILarksuiteAppointment, lead?: IFrappeLead) {
-    const payload = await this.mapLarkToFrappe(dataRequest, lead);
+    const payload = await mapLarkToFrappe(this.frappeClient, dataRequest, lead);
     const data = await this.frappeClient.insert({
       doctype: "Appointment",
       ...payload
@@ -250,7 +161,7 @@ export default class AppointmentService {
 
     if (existing && existing.length > 0) {
       const docName = existing[0].name;
-      const payload = await this.mapLarkToFrappe(dataRequest, lead);
+      const payload = await mapLarkToFrappe(this.frappeClient, dataRequest, lead);
       const data = await this.frappeClient.update({
         doctype: "Appointment",
         name: docName,
